Fixed #5 Improved test to assure type comparison takes place

+++ src/WebGLDotNET/WebGL.cs
@@ -789,7 +789,7 @@ namespace WebGLDotNET
 
         public object GetParameter(uint pname) => Invoke("getParameter", pname);
 
-        public uint GetError() => InvokeForBasicType<uint>("getError");
+        public uint GetError() => (uint)InvokeForBasicType<int>("getError");
 
         public object GetFramebufferAttachmentParameter(uint target, uint attachment, uint pname) => Invoke("getFramebufferAttachmentParameter", target, attachment, pname);
